+++ CHANGELOG.md
@@ -29,6 +29,7 @@
  - Wand haste, hunger and health regen use potion effects (looks better, avoids conflicts)
  - Curse, Boon, Wither etc effects stack properly
  - Rewind and Rollback will show the name of the spell they undid
+ - Add Mount spell
 
 ## 3.0